+++ projects/page-builder/src/lib/block/block.component.ts
@@ -3,23 +3,13 @@ import {
   ChangeDetectorRef,
   Component,
   EventEmitter,
-  Inject,
   Injector,
   Input,
   OnDestroy,
-  Optional,
   Output,
   ViewChild
 } from '@angular/core';
-import {
-  CUSTOM_FIELDS,
-  CustomFields,
-  FormBuilderComponent,
-  FormBuilderContextService,
-  FormBuilderData,
-  Parser,
-  State
-} from '@jaspero/form-builder';
+import {FormBuilderComponent, FormBuilderData, Parser, State} from '@jaspero/form-builder';
 import {Subscription} from 'rxjs';
 import {Selected} from '../selected.interface';
 
@@ -32,11 +22,7 @@ import {Selected} from '../selected.interface';
 export class BlockComponent implements OnDestroy {
   constructor(
     private injector: Injector,
-    @Optional()
-    @Inject(CUSTOM_FIELDS)
-    private customFields: CustomFields,
-    private cdr: ChangeDetectorRef,
-    private ctx: FormBuilderContextService
+    private cdr: ChangeDetectorRef
   ) {}
 
   @Input() parentFormId = 'main';
@@ -66,11 +52,7 @@ export class BlockComponent implements OnDestroy {
       // TODO: Replace with correct state
       State.Create,
       'admin',
-      selected.form.definitions,
-      {
-        ...this.customFields || {},
-        ...this.ctx.fields
-      }
+      selected.form.definitions
     );
 
     this.parser.form = this.parser.buildForm(