+++ src/app/examples/core/core.module.ts
@@ -1,9 +1,8 @@
 import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import {FormBuilderModule} from '../../../../projects/form-builder/src/lib/form-builder.module';
+import {FormBuilderModule} from '@jaspero/form-builder';
 import {CoreComponent} from './core.component';
-import {FormBuilderContextService} from '../../../../projects/form-builder/src/public-api';
 
 const routes: Routes = [{
   path: '',
@@ -17,9 +16,6 @@ const routes: Routes = [{
     RouterModule.forChild(routes),
 
     FormBuilderModule
-  ],
-  providers: [
-    FormBuilderContextService
   ]
 })
 export class CoreModule { }