+++ src/app/examples/core/forms/form-ui-example.ts
@@ -1,9 +1,8 @@
-import {SegmentType} from '../../../../../projects/form-builder/src/lib/enums/segment-type.enum';
 import {FormBuilderData} from '../../../../../projects/form-builder/src/lib/interfaces/form-builder-data.interface';
 
 export const FORM_UI_EXAMPLE: FormBuilderData = {
   segments: [{
-    type: SegmentType.Empty,
+    type: 'empty',
     fields: ['/fields']
   }],
   schema: {