+++ projects/form-ui/CHANGELOG.md
@@ -1,3 +1,32 @@
+## @jaspero/fb-form-ui [2.3.3](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@2.3.2...@jaspero/fb-form-ui@2.3.3) (2021-11-08)
+
+
+### Bug Fixes
+
+* **form-ui:** removed SegmentType references ([c519f18](https://github.com/Jaspero/schema-forms/commit/c519f18f740055b2d87889613d924c8997f60556))
+
+## @jaspero/fb-form-ui [2.3.2](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@2.3.1...@jaspero/fb-form-ui@2.3.2) (2021-11-08)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 4.18.2
+* **@jaspero/fb-tinymce:** upgraded to 2.2.42
+
+## @jaspero/fb-form-ui [2.3.1](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@2.3.0...@jaspero/fb-form-ui@2.3.1) (2021-11-08)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 4.18.1
+* **@jaspero/fb-tinymce:** upgraded to 2.2.41
+
 # @jaspero/fb-form-ui [2.3.0](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@2.2.39...@jaspero/fb-form-ui@2.3.0) (2021-11-07)