+++ projects/segments-mat/CHANGELOG.md
@@ -1,3 +1,23 @@
+## @jaspero/fb-segments-mat [1.0.2](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-segments-mat@1.0.1...@jaspero/fb-segments-mat@1.0.2) (2021-11-08)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 4.18.2
+
+## @jaspero/fb-segments-mat [1.0.1](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-segments-mat@1.0.0...@jaspero/fb-segments-mat@1.0.1) (2021-11-08)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 4.18.1
+
 # @jaspero/fb-segments-mat 1.0.0 (2021-11-07)